feat:增加权限检查服务和用户权限表
修改数据库：增加用户权限表
增加1个实体
增加1个repository
修改1个repository：修改数据类型
修改2个实体：修改数据类型和字段名
增加1个服务：权限检查方法

+++ src/main/java/com/example/demo/entity/Role.java
@@ -7,7 +7,6 @@
 
 package com.example.demo.entity;
 
-import java.math.BigInteger;
 import javax.persistence.*;
 
 import lombok.*;
@@ -51,5 +50,5 @@ public class Role
             nullable = false,
             updatable = false
     )
-    private BigInteger permissions;
+    private Long permissions;
 }
